% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_gibbs.R
\name{breakpoint.gibbs}
\alias{breakpoint.gibbs}
\title{Gibbs-Sampler für Poisson-Bruchpunktmodell}
\usage{
breakpoint.gibbs(size, y)
}
\arguments{
\item{size}{Anzahl Iterationen}

\item{y}{Daten}
}
\value{
Ziehungen aus der Posteriori von (lambda, alpha, theta)
}
\description{
Gibbs-Sampler für Poisson-Bruchpunktmodell
}
\author{
Michael Höhle (?)
}
