% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsBLasso.R
\name{gibbsBLasso}
\alias{gibbsBLasso}
\title{Gibbs Sampler for Bayesian Lasso}
\source{
https://cs.gmu.edu/~pwang7/gibbsBLasso.html
}
\usage{
gibbsBLasso(x, y, max.steps = 1e+05)
}
\arguments{
\item{x}{Desing matrix}

\item{y}{Data}

\item{max.steps}{Number of iterations}
}
\value{
Vector of posterior median beta
}
\description{
Gibbs Sampler for Bayesian Lasso
}
\examples{
library(lars)
data(diabetes)
x <- scale(diabetes$x)
y <- scale(diabetes$y)
betas <- gibbsBLasso(x, y, max.steps = 100000)
}
\author{
Pu Wang, George Mason University, Fairfax, USA
}
