% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson-lognormal.R
\name{poisson.lognormal.mcmc}
\alias{poisson.lognormal.mcmc}
\title{MCMC Beispiel für Poisson-Likelihood mit Lognormal-Priori}
\usage{
poisson.lognormal.mcmc(sumx, n, prior = list(a = 1, b = 1, tau = 1000),
  I = 1000, do.tuning = FALSE, tuning = 50, do.burnin = FALSE,
  burnin = 100)
}
\arguments{
\item{sumx}{Summe der Daten}

\item{n}{Anzahl Daten}

\item{prior}{Liste mit Priori-Parametern, default: list("a"=1,"b"=1,"tau"=1000)}

\item{I}{Anzahl Iterationen}

\item{do.tuning}{Soll getunt werden?}

\item{tuning}{Anzahl Tuning-Iterationen}

\item{do.burnin}{Soll burnin beachtet werden}

\item{burnin}{Anzahl burnin-Iterationen}
}
\value{
coda::mcmc Objekt
}
\description{
MCMC Beispiel für Poisson-Likelihood mit Lognormal-Priori
}
