% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{rmvnormcanon}
\alias{rmvnormcanon}
\title{Draw random vectors from multivariate Gaussian in canonical form}
\usage{
rmvnormcanon(b, P)
}
\arguments{
\item{b}{b parameter (canonical mean)}

\item{P}{Precision matrix}
}
\value{
Vector of draw from N(P^{-1}b,P^{-1})
}
\description{
Draw random vectors from multivariate Gaussian in canonical form
}
\examples{
rmvnormcanon(c(0,0),diag(2))
}
